refactor: Container와 Injector 로직 분리
