+++ app/src/main/java/woowacourse/shopping/ui/ViewModelFactory.kt
@@ -2,10 +2,10 @@ package woowacourse.shopping.ui
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import com.woowacourse.di.DependencyContainer
+import com.woowacourse.di.Injector
 
 class ViewModelFactory : ViewModelProvider.Factory {
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        return DependencyContainer.createInstance(modelClass.kotlin)
+        return Injector.createInstance(modelClass.kotlin)
     }
 }
