+++ di/src/main/java/com/woowacourse/di/DiActivity.kt
@@ -1,17 +1,15 @@
 package com.woowacourse.di
 
 import android.os.Bundle
-import android.util.Log
 import androidx.appcompat.app.AppCompatActivity
 
 abstract class DiActivity : AppCompatActivity() {
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setupModule()
     }
 
     private fun setupModule() {
-        DependencyContainer.injectProperty(this)
+        Injector.injectProperty(this)
     }
 }
